Move SUSE exception classes to separate file
